+++ core/server/api_container/server/service_network/service_network.go
@@ -229,7 +229,6 @@ func (network *ServiceNetwork) StartService(
 	filesArtifactMountDirpaths map[service.FilesArtifactID]string,
 ) (
 	resultServiceGuid service.ServiceGUID,
-	resultPrivateIpAddr net.IP,
 	resultMaybePublicIpAddr net.IP, // Will be nil if the service doesn't declare any private ports
 	resultPublicPorts map[string]*port_spec.PortSpec,
 	resultErr error,
@@ -240,7 +239,7 @@ func (network *ServiceNetwork) StartService(
 
 	registration, found := network.registeredServiceInfo[serviceId]
 	if !found {
-		return "", nil, nil, nil, stacktrace.NewError("Cannot start service; no registration exists for service with ID '%v'", serviceId)
+		return "", nil, nil, stacktrace.NewError("Cannot start service; no registration exists for service with ID '%v'", serviceId)
 	}
 	serviceGuid := registration.GetGUID()
 
@@ -254,7 +253,7 @@ func (network *ServiceNetwork) StartService(
 	if network.isPartitioningEnabled {
 		servicePacketLossConfigurationsByServiceID, err := network.topology.GetServicePacketLossConfigurationsByServiceID()
 		if err != nil {
-			return "", nil, nil, nil, stacktrace.Propagate(err, "An error occurred getting the packet loss configuration by service ID "+
+			return "", nil, nil, stacktrace.Propagate(err, "An error occurred getting the packet loss configuration by service ID "+
 				" to know what packet loss updates to apply on the new node")
 		}
 
@@ -272,7 +271,7 @@ func (network *ServiceNetwork) StartService(
 			network.registeredServiceInfo,
 			network.networkingSidecars,
 		); err != nil {
-			return "", nil, nil, nil, stacktrace.Propagate(
+			return "", nil, nil, stacktrace.Propagate(
 				err,
 				"An error occurred updating the traffic control configuration of all the other services "+
 					 "before adding the new service, meaning that the service wouldn't actually start in a partition",
@@ -292,7 +291,7 @@ func (network *ServiceNetwork) StartService(
 		filesArtifactMountDirpaths,
 	)
 	if err != nil {
-		return "", nil, nil, nil, stacktrace.Propagate(
+		return "", nil, nil, stacktrace.Propagate(
 			err,
 			"An error occurred starting service '%v'",
 			serviceId,
@@ -303,19 +302,19 @@ func (network *ServiceNetwork) StartService(
 	if network.isPartitioningEnabled {
 		sidecar, err := network.networkingSidecarManager.Add(ctx, registration.GetGUID())
 		if err != nil {
-			return "", nil, nil, nil, stacktrace.Propagate(err, "An error occurred adding the networking sidecar")
+			return "", nil, nil, stacktrace.Propagate(err, "An error occurred adding the networking sidecar")
 		}
 		network.networkingSidecars[serviceId] = sidecar
 
 		if err := sidecar.InitializeTrafficControl(ctx); err != nil {
-			return "", nil, nil, nil, stacktrace.Propagate(err, "An error occurred initializing the newly-created networking-sidecar-traffic-control-qdisc-configuration")
+			return "", nil, nil, stacktrace.Propagate(err, "An error occurred initializing the newly-created networking-sidecar-traffic-control-qdisc-configuration")
 		}
 
 		// TODO Getting packet loss configuration by service ID is an expensive call and, as of 2021-11-23, we do it twice - the solution is to make
 		//  Getting packet loss configuration by service ID not an expensive call
 		servicePacketLossConfigurationsByServiceID, err := network.topology.GetServicePacketLossConfigurationsByServiceID()
 		if err != nil {
-			return "", nil, nil, nil, stacktrace.Propagate(err, "An error occurred getting the packet loss configuration by service ID "+
+			return "", nil, nil, stacktrace.Propagate(err, "An error occurred getting the packet loss configuration by service ID "+
 				" to know what packet loss updates to apply on the new node")
 		}
 		newNodeServicePacketLossConfiguration := servicePacketLossConfigurationsByServiceID[serviceId]
@@ -323,12 +322,12 @@ func (network *ServiceNetwork) StartService(
 			serviceId: newNodeServicePacketLossConfiguration,
 		}
 		if err := updateTrafficControlConfiguration(ctx, updatesToApply, network.registeredServiceInfo, network.networkingSidecars); err != nil {
-			return "", nil, nil, nil, stacktrace.Propagate(err, "An error occurred applying the traffic control configuration on the new node to partition it "+
+			return "", nil, nil, stacktrace.Propagate(err, "An error occurred applying the traffic control configuration on the new node to partition it "+
 				"off from other nodes")
 		}
 	}
 
-	return serviceGuid, registration.GetPrivateIP(), userService.GetMaybePublicIP(), userService.GetMaybePublicPorts(), nil
+	return serviceGuid, userService.GetMaybePublicIP(), userService.GetMaybePublicPorts(), nil
 }
 
 func (network *ServiceNetwork) RemoveService(